feat(dap): Add fast crc32 verify algorithm
Different devices may have different levels of permissions. In some
cases, for example, the contents of the ROM cannot be accessed through
the debug path.

An interesting phenomenon occurs when the "Init()" function in the
flash algorithm is executed, the ROM may enter a special state
where the CPU itself cannot access the contents of the flash, but
it can be accessed through debug path.

Therefore, this patch removes the previously executed "Init" and
"Uninit" functions during the verification process. Setting aside
certain security considerations, when these functions are not executed,
the CPU should be able to access the contents of the flash memory.

There might be cases where the flash memory is set to execute only,
perhaps at that point we should consider adopting one of these methods,
or just use the verify method that comes with the flash algorithm.